Update transform value

+++ shared/dto/set.balance.dto.ts
@@ -12,7 +12,7 @@ export class SetBalanceDTO {
     type: Number,
     required: true
   })
-  @Transform(({ value: amount }) => parseInt(amount))
+  @Transform(({ value: amount }) => parseFloat(amount))
   @IsPositive()
   @IsNumber()
   amount: number;
