+++ shared/dto/set.obligation.dto.ts
@@ -27,7 +27,7 @@ export class SetObligationDTO {
     type: Number,
     required: true
   })
-  @Transform(({ value: percent }) => parseInt(percent))
+  @Transform(({ value: percent }) => parseFloat(percent))
   @IsPositive()
   @IsNumber()
   percent: number;
@@ -36,7 +36,7 @@ export class SetObligationDTO {
     type: Number,
     required: true
   })
-  @Transform(({ value: insurance }) => parseInt(insurance))
+  @Transform(({ value: insurance }) => parseFloat(insurance))
   @IsPositive()
   @IsNumber()
   insurance: number;
@@ -45,7 +45,7 @@ export class SetObligationDTO {
     type: Number,
     required: true
   })
-  @Transform(({ value: current_amount }) => parseInt(current_amount))
+  @Transform(({ value: current_amount }) => parseFloat(current_amount))
   @IsPositive()
   @IsNumber()
   current_amount: number;
@@ -54,7 +54,7 @@ export class SetObligationDTO {
     type: Number,
     required: true
   })
-  @Transform(({ value: total_amount }) => parseInt(total_amount))
+  @Transform(({ value: total_amount }) => parseFloat(total_amount))
   @IsPositive()
   @IsNumber()
   total_amount: number;
