+++ shared/dto/update.balance.dto.ts
@@ -20,7 +20,7 @@ export class UpdateBalanceDTO {
     type: Number,
     required: true
   })
-  @Transform(({ value: amount }) => parseInt(amount))
+  @Transform(({ value: amount }) => parseFloat(amount))
   @IsPositive()
   @IsNumber()
   amount: number;
