+++ shared/dto/update.operation.dto.ts
@@ -41,7 +41,7 @@ export class UpdateOperationDTO {
     type: Number,
     required: true
   })
-  @Transform(({ value: amount }) => parseInt(amount))
+  @Transform(({ value: amount }) => parseFloat(amount))
   @IsPositive()
   @IsNumber()
   amount: number;
