+++ shared/dto/update.operation.type.dto.ts
@@ -29,7 +29,7 @@ export class UpdateOperationTypeDTO {
     type: Number,
     required: true
   })
-  @Transform(({ value: commission }) => parseInt(commission))
+  @Transform(({ value: commission }) => parseFloat(commission))
   @IsPositive()
   @IsNumber()
   commission: number;
